Factor out resultsTable
